requirements.txt
# QBN (QuantBayes Nexus) Dependencies
# =====================================
# GPU container met CUDA 12.x support

# === Core Data Science ===
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
opt_einsum>=3.3.0  # Faster pgmpy inference

# === Database ===
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
asyncpg>=0.28.0  # Async database support
python-dotenv>=1.0.0
pyyaml>=6.0.1

# === Bayesian Network (CPU - pgmpy) ===
pgmpy>=0.1.23
networkx>=3.1

# === GPU Computing ===
cupy-cuda12x>=12.0.0  # GPU-accelerated NumPy
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# === Bayesian Neural Networks ===
torchbnn>=1.2  # Bayesian Linear layers

# === Progress & Logging ===
tqdm>=4.66.0
colorlog>=6.7.0
rich>=13.0.0  # Terminal progress bars

# === Visualization (Threshold Optimizer) ===
matplotlib>=3.7.0
seaborn>=0.12.0

# === API (toekomstig) ===
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# === Development & Testing ===
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
ipython>=8.0.0

